Finish with type inference
